A good few hours of work learning some JavaScript.
The objective was to create some JS functions that generated a time line
reversed calendar for each year, with a starting year to the current year,
with the months also reversed.

Instead of having a 'list' of calendars from the starting year to the
current being displayed/added to the document in that order (2006 ->
<current-year>), the order is reversed, such that the <current-year> is
the first year displayed under the button.

The months are similarly featured in reversed (descending) order.

Could this be easier? Probably.

Am I going to learn <insert random flavor of a JS/CSS framework/library
combination that is already outdated by the time it is conceived .. here>?

Nope. That isn't how I learn the basics. That's what this is. If you are
asking the question above, you're probably not going to learn anything here. :)