Renames the testing file to an actual thing file.
The repository was intially meant to be a sandbox, for testing things
out. However, upon further thinking, I decided to upgrade it into a
project-type repository, and therefore, it is actually a thing now.

The contents of the repository should therefore be named accordingly,
so the first step is to rename and reoganize. This commit renames the
HTML file to be the index page.