Adds actual resume data for elementary descriptions display.
I was starting to add some boring "Lorem Ipsum" kind of text, and
as I was setting up the data, I decided to just pull in the data I
am using in a Python+Flask experiement page, and therefore it is
no longer "lorem" descriptions, but "actual-headings".

There are additional details to be exploited in the near future,
but the objective was to connect reading data (from a static JSON
file) to displaying some of that data in the descriptions section.

It is not displayed well (black text on green background), and the
updated snapshot shows that.