Adds the small detail of linking to an organization/institution.
A lot of renames/moving of files/functions, as I was starting to get lost on my
little 768p laptop monitor in VS Code, but for real, I like modularity in my code,
not "one file to rule them all", even if there is a bunch of copy-pasta-different-
sauce everywhere on the table. The corresponding data was split as well; and I
finally moved timeline.js into the code section of the static files.

Finally, removed the Add Descriptions button, as the functionality has been moved
into the onPageLoaded function called when the page is loaded.