Adds location to the organization/institution line.
Well, the RGB-ness has disappeared, and I added headers for the tags column,
in anticipation of each description having a tag that is categorized with
such headers.

I have also replaced the unordered list with sections and articles (article.css and articulated.js)
which allowed me to remove a lot of the copy-pasta'd generation of description entries.

Finally, I played around with the colors, for the sections/headers, and as well as article title
links when hovered over by the cursor.