Adds the display of tags to their respective groups.
The "Methodologies" section was renamed on a static variant of my one-page
resume to "Processes", and this is now incorporated into the interactive
version.

There's quite a bit of styling to decide and apply, but for now, the desired
functionality is available. In the future, the tag data won't be coming from
tags.json, but from experience.json and educations.json.